+++ mountwizzard/environment/unihedron_thread.py
@@ -85,7 +85,6 @@ class Unihedron(PyQt5.QtCore.QThread):
                 finally:                                                                                                    # still continua and try it again
                     pass                                                                                                    # needed for continue
                 time.sleep(5)                                                                                               # wait for the next cycle
-        self.ascom.Quit()
         pythoncom.CoUninitialize()                                                                                          # needed for doing COm objects in threads
         self.terminate()                                                                                                    # closing the thread at the end